Refactor spec discovery into focused modules
