+++ packages/analyzer/src/spec-discovery/index.ts
@@ -1,0 +1,4 @@
+export { SPEC_PROSE_EXTRACTOR, SPEC_STRUCTURED_EXTRACTOR } from './constants.js'
+export { discoverSpecFiles } from './discovery.js'
+export { createSpecExtractionManifest } from './manifest.js'
+export { parseMarkdownSpec, parseSpecContent, parseTextSpec } from './prose.js'
